Refactor executor tests into testkit.
