+++ mleap-executor-testkit/build.sbt
@@ -1,0 +1,4 @@
+import ml.combust.mleap.{Dependencies, Common}
+
+Common.defaultMleapServingSettings
+Dependencies.executorTestKit
