+++ mleap-executor-testkit/src/main/resources/leap_frame/frame.airbnb.avro
