+++ mleap-executor-tests/build.sbt
@@ -1,0 +1,3 @@
+import ml.combust.mleap.{Dependencies, Common}
+
+Common.defaultMleapServingSettings
